# packages/contracts/openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: SuperAssistant API
  version: 1.1.0
  description: MVP v1 baseline APIs with core admin endpoints.
servers:
  - url: /api/v1
components:
  parameters:
    XRequestId:
      in: header
      name: X-Request-Id
      required: true
      schema:
        type: string
    XAppVersion:
      in: header
      name: X-App-Version
      required: true
      schema:
        type: string
    XPlatform:
      in: header
      name: X-Platform
      required: true
      schema:
        type: string
        enum: [ios, android, web-admin]
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /categories:
    get:
      summary: List custom and system categories
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
      responses:
        '200': { description: success }
    post:
      summary: Create a custom category
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, unit_type]
              properties:
                name: { type: string, minLength: 1, maxLength: 32 }
                icon: { type: string, maxLength: 64, nullable: true }
                unit_type: { type: string, minLength: 1, maxLength: 16 }
                decay_template: { type: object, nullable: true, additionalProperties: true }
      responses:
        '200': { description: success }
        '409': { description: conflict }
  /categories/{category_id}:
    patch:
      summary: Update a custom category
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: category_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 32 }
                icon: { type: string, maxLength: 64, nullable: true }
                unit_type: { type: string, minLength: 1, maxLength: 16 }
                decay_template: { type: object, nullable: true, additionalProperties: true }
      responses:
        '200': { description: success }
        '403': { description: forbidden (system category) }
        '404': { description: not found }
    delete:
      summary: Delete a custom category
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: category_id
          required: true
          schema: { type: string }
      responses:
        '200': { description: success }
        '403': { description: forbidden (system category) }
        '404': { description: not found }
        '409': { description: conflict (category has activity history) }
  /households/me:
    get:
      summary: Get current household
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
      responses:
        '200': { description: success }
        '404': { description: not found }
  /households:
    post:
      summary: Create a household
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 64 }
      responses:
        '200': { description: success }
        '409': { description: conflict (must leave current household first) }
  /households/invitations:
    post:
      summary: Create a household invitation
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [ADMIN, MEMBER, VIEWER], default: MEMBER }
      responses:
        '200': { description: success }
        '403': { description: forbidden (OWNER/ADMIN role required) }
  /households/join:
    post:
      summary: Join a household by invite code
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code: { type: string, minLength: 4, maxLength: 16 }
      responses:
        '200': { description: success }
        '404': { description: not found (invalid invite code) }
        '410': { description: gone (invitation expired) }
  /inventory/items:
    get:
      summary: Get inventory list
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - in: query
          name: status
          schema:
            type: string
            enum: [PLENTY, WARNING, CRITICAL, EMPTY]
      responses:
        '200': { description: success }
  /inventory/operations/batch:
    post:
      summary: Batch inventory operations
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [operations]
              properties:
                operations:
                  type: array
                  items:
                    type: object
                    required: [item_key, operation, value, unit, source]
                    properties:
                      item_key: { type: string }
                      operation: { type: string, enum: [ADD, SET, SUBTRACT, CLEAR] }
                      value: { type: number }
                      unit: { type: string }
                      source: { type: string, enum: [VOICE, OCR, MANUAL] }
                      op_id: { type: string }
                      client_version: { type: integer }
      responses:
        '200': { description: success }
  /inventory/offline/replay:
    post:
      summary: Enqueue offline operations for replay
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [offline_ops]
              properties:
                offline_ops:
                  type: array
                  items:
                    type: object
                    required: [op_id, payload]
                    properties:
                      op_id: { type: string }
                      created_at: { type: string, format: date-time }
                      retry_count: { type: integer, minimum: 0 }
                      payload:
                        type: object
                        required: [operations]
                        properties:
                          operations:
                            type: array
                            items:
                              type: object
                              required: [item_key, operation, value, unit, source]
                              properties:
                                item_key: { type: string }
                                operation: { type: string, enum: [ADD, SET, SUBTRACT, CLEAR] }
                                value: { type: number }
                                unit: { type: string }
                                source: { type: string, enum: [VOICE, OCR, MANUAL] }
                                op_id: { type: string }
                                client_version: { type: integer }
      responses:
        '200': { description: success }
  /inventory/logs:
    get:
      summary: Get inventory logs
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - in: query
          name: page
          schema: { type: integer, minimum: 1 }
        - in: query
          name: page_size
          schema: { type: integer, minimum: 1, maximum: 200 }
        - in: query
          name: item_key
          schema: { type: string }
      responses:
        '200': { description: success }
  /inventory/items/{item_key}/calibrate:
    post:
      summary: Calibrate one inventory item
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/XRequestId'
        - $ref: '#/components/parameters/XAppVersion'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/IdempotencyKey'
        - in: path
          name: item_key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, unit]
              properties:
                value: { type: number }
                unit: { type: string }
      responses:
        '200': { description: success }
...